requirements.txt
openai==1.61.1
python-dotenv==1.0.1
requests==2.32.3